Moved ConditionalRandomField to new module
Renamed module allennlp.modules.conditional_random_field_weight
to ...conditional_random_files
